Extract `CommandDeduplicationAssertions` from `Assertions`, change `assertCompletionIsDefined` to `assertDefined`
CHANGELOG_BEGIN
CHANGELOG_END

run-full-compat: true
