+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/suites/CompletionDeduplicationInfoIT.scala
@@ -5,7 +5,7 @@ package com.daml.ledger.api.testtool.suites
 
 import com.daml.ledger.api.SubmissionIdGenerator
 import com.daml.ledger.api.testtool.infrastructure.Allocation._
-import com.daml.ledger.api.testtool.infrastructure.Assertions.assertCompletionIsDefined
+import com.daml.ledger.api.testtool.infrastructure.Assertions.assertDefined
 import com.daml.ledger.api.testtool.infrastructure.LedgerTestSuite
 import com.daml.ledger.api.testtool.infrastructure.participant.ParticipantTestContext
 import com.daml.ledger.api.testtool.suites.CompletionDeduplicationInfoIT._
@@ -101,7 +101,7 @@ private[testtool] object CompletionDeduplicationInfoIT {
       } yield completion
 
     override def assertCompletion(optCompletion: Option[Completion]): Unit = {
-      val completion = assertCompletionIsDefined(optCompletion)
+      val completion = assertDefined(optCompletion, "No completion has been produced")
       assert(completion.status.forall(_.code == Status.Code.OK.value()))
       assert(
         Ref.SubmissionId.fromString(completion.submissionId).isRight,
@@ -136,7 +136,7 @@ private[testtool] object CompletionDeduplicationInfoIT {
       } yield completion
 
     override def assertCompletion(optCompletion: Option[Completion]): Unit = {
-      val completion = assertCompletionIsDefined(optCompletion)
+      val completion = assertDefined(optCompletion, "No completion has been produced")
       assert(completion.status.forall(_.code == Status.Code.OK.value()))
     }
   }
@@ -156,7 +156,7 @@ private[testtool] object CompletionDeduplicationInfoIT {
       optCompletion: Option[Completion],
       requestedSubmissionId: Ref.SubmissionId,
   ): Unit = {
-    val submissionIdCompletion = assertCompletionIsDefined(optCompletion)
+    val submissionIdCompletion = assertDefined(optCompletion, "No completion has been produced")
     val actualSubmissionId = submissionIdCompletion.submissionId
     assert(submissionIdCompletion.status.forall(_.code == Status.Code.OK.value()))
     assert(
@@ -169,7 +169,7 @@ private[testtool] object CompletionDeduplicationInfoIT {
   private def assertDeduplicationPeriodIsReported(
       optCompletion: Option[Completion]
   ): Unit = {
-    val completion = assertCompletionIsDefined(optCompletion)
+    val completion = assertDefined(optCompletion, "No completion has been produced")
     assert(completion.status.forall(_.code == Status.Code.OK.value()))
     assert(completion.deduplicationPeriod.isDefined, "The deduplication period was not reported")
   }
@@ -179,7 +179,7 @@ private[testtool] object CompletionDeduplicationInfoIT {
       optCompletion: Option[Completion],
   ): Unit = {
     val expectedApplicationId = requestedApplicationId
-    assertCompletionIsDefined(optCompletion)
+    assertDefined(optCompletion, "No completion has been produced")
     val applicationIdCompletion = optCompletion.get
     assert(applicationIdCompletion.status.forall(_.code == Status.Code.OK.value()))
     val actualApplicationId = applicationIdCompletion.applicationId
